requirements.txt
langchain
langgraph
langchain-openai
langchain-community
langchain-chroma
langchain-ollama
langchain-google-genai
langchain-groq
gradio
python-dotenv
bs4
wikipedia
ipykernel
pypdf
sentence-transformers